+++ changelog.md
@@ -3,6 +3,7 @@
 - progress rool actions
 - error screen
 - about page
+- delete challenge
 
 # 0.1.0